feat: Refactor NumberFormatMixin to helper class with static methods, update PageDetailsBlock and BlockDetailExt to use new number formatting
